requirements.txt
## The following requirements were added by pip --freeze:
Django==1.5
South==0.7.6
Unidecode==0.04.12
decorator==3.4.0
django-extensions==1.1.1
django-facebook==5.0.12
django-picklefield==0.3.0
django-tastypie==0.9.14
django-watson==1.1.1
facebook-sdk==0.4.0
google-api-python-client==1.1
httplib2==0.8
lucene-querybuilder==0.1.6
mimeparse==0.1.3
musicbrainzngs==0.3
neo4django==0.1.8
neo4jrestclient==1.8.0
psycopg2==2.5
python-dateutil==2.1
python-gflags==2.0
requests==1.2.0
six==1.3.0
wsgiref==0.1.2
